% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_Comparison.R
\name{Coereba_Comparison}
\alias{Coereba_Comparison}
\title{Compares two Coereba populations to find difference,
returning either data or a plot of the marker expressions}
\usage{
Coereba_Comparison(gs, gs2 = NULL, Arg1, Arg2, panel, returnType)
}
\arguments{
\item{gs}{The GatingSet}

\item{gs2}{Default NULL, if comparing a different GatingSet}

\item{Arg1}{First Population Name}

\item{Arg2}{Second Population Name, can be identical if different GatingSet}

\item{panel}{A data.frame or a file.path to the Panel .csv}

\item{returnType}{Whether to return "plot" or "data". Data default.}
}
\value{
Either data or a ggplot2 object comparing the two populations of interest
}
\description{
Compares two Coereba populations to find difference,
returning either data or a plot of the marker expressions
}
\examples{

library(Coereba)
}
