% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_FCS.R
\name{Coereba_FCS_Reversal}
\alias{Coereba_FCS_Reversal}
\title{Handles Description Digging for Dictionary Reversal going FCS to data}
\usage{
Coereba_FCS_Reversal(Coereba)
}
\arguments{
\item{Coereba}{Either a flow/cytoframe with embedded Keywords. Alternately, the file.path to the .fcs file
with the same embedded keywords}
}
\value{
A data.frame containing the reversed columns to be passed on to SummarizedExperiment
}
\description{
Handles Description Digging for Dictionary Reversal going FCS to data
}
\examples{
A <- 2 + 2
}
