% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_FCS.R
\name{Coereba_FCSExport}
\alias{Coereba_FCSExport}
\title{Takes Coereba Output and generates both .fcs file and a dictionary.}
\usage{
Coereba_FCSExport(
  data,
  gs,
  outpath,
  filename,
  fcsname,
  returnType,
  nameAppend,
  Aggregate = FALSE,
  coerebaCombine = FALSE
)
}
\arguments{
\item{data}{The Coereba Data}

\item{gs}{Gating Set Object containing original parameters}

\item{outpath}{File path to desired storage location}

\item{filename}{Desired name for fcs file}

\item{returnType}{Whether to return "flowframe" or "fcs"}

\item{nameAppend}{For flowframe and fcs returnType, what gets appended before .fcs}

\item{Aggregate}{Combine the items, finish your documentation.}

\item{combineCoereba}{Internal, used for Coereba flowframe aggregation}
}
\value{
A dictionary .csv file, and a .fcs file.
}
\description{
Takes Coereba Output and generates both .fcs file and a dictionary.
}
\examples{
A <- 2+2
}
