% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_MarkerExpressions.R
\name{Coereba_MarkerExpressions}
\alias{Coereba_MarkerExpressions}
\title{Generate overall marker expressions across Coereba clusters,}
\usage{
Coereba_MarkerExpressions(
  x,
  theassay = "ratios",
  returnType = "All",
  CombinatorialArgs = NULL
)
}
\arguments{
\item{x}{The Coereba Summarized Experiment object from Coereba_Processing}

\item{returnType}{Either "All" or "Combinatorial". Default is ALL.}

\item{CombinatorialArgs}{When returnType Combinatorial, the two fluorophores to create quadrants for.}

\item{TheAssay}{Default is "ratios"}
}
\value{
returns a data.frame
}
\description{
Generate overall marker expressions across Coereba clusters,
}
\examples{

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")

files <- gs_pop_get_data(gs, "CD4")
MySE <- Coereba_Processing(x=files[[1]], panel=panelPath)

All <- Coereba_MarkerExpressions2(x=MySE)

Memory <- Coereba_MarkerExpressions2(x=MySE, returnType = "Combinatorial",
 CombinatorialArgs=c("BV510", "APC-Fire 750"))

}
