% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_UpdateGates.R
\name{Coereba_UpdateGates}
\alias{Coereba_UpdateGates}
\title{Update a gate cutoff csv with the Coereba_App edits}
\usage{
Coereba_UpdateGates(Clicks, Old, export = TRUE, outpath, fileName)
}
\arguments{
\item{Clicks}{A path to a .csv file or a folder containing only the Click .csv files.}

\item{Old}{File path to the CSV containing the original Estimated Gate Cutoffs}

\item{export}{Whether to export the updated .csv, default is TRUE}

\item{outpath}{File path to save new .csv file at.}

\item{fileName}{Desired name for the updated .csv file}
}
\value{
An updated .csv file to the new location, and a data.frame
}
\description{
Update a gate cutoff csv with the Coereba_App edits
}
\examples{

File_Location <- system.file("extdata", package = "Coereba")
TheOldCSV <- file.path(File_Location, "GateCutoffsForNKs.csv")
TheClickInfo <- file.path(File_Location, "ClickDataExample.csv")

UpdatedCSV <- Coereba_UpdateGates(Clicks=TheClickInfo, Old=TheOldCSV,
 export=FALSE, outpath=NULL, fileName="UpdatedCSV")

}
