% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModernGateCutoffs.R
\name{ColumnExprs}
\alias{ColumnExprs}
\title{Internal for Modern Gate Cutoff}
\usage{
ColumnExprs(x, TheDF, w, span, ...)
}
\arguments{
\item{x}{Something}

\item{TheDF}{Something}

\item{w}{Something}

\item{span}{Something}

\item{...}{Something}
}
\description{
Internal for Modern Gate Cutoff
}
\keyword{internal}
