% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complicated.R
\name{Complicated}
\alias{Complicated}
\title{Original Reach Implementation (superceeded)}
\usage{
Complicated(x, cutoff, starter)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{cutoff}{Number of cells required to retain the cluster}

\item{starter}{A column to start division on, (ex. "BV711-A")}
}
\description{
Original Reach Implementation (superceeded)
}
\examples{
NULL
}
