% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytokineCleanup.R
\name{CytokineCleanup}
\alias{CytokineCleanup}
\title{Combine FlowJo Q1 Q2 columns}
\usage{
CytokineCleanup(x, variable, first, second)
}
\arguments{
\item{x}{A data.frame object}

\item{variable}{A column name}

\item{first}{Return name first column}

\item{second}{Return name second column}
}
\description{
Combine FlowJo Q1 Q2 columns
}
\examples{
NULL
}
