% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagoTest.R
\name{dagoTest}
\alias{dagoTest}
\title{Dagostino Omnibus K2 wrapper adapted from fBasics}
\usage{
dagoTest(x, title = NULL, description = NULL)
}
\arguments{
\item{x}{Some Text}

\item{title}{Some Text}

\item{description}{Some Text}
}
\value{
Some Text
}
\description{
Dagostino Omnibus K2 wrapper adapted from fBasics
}
\examples{
NULL

}
