% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_MarkerExpressions.R
\name{DataRetrieval}
\alias{DataRetrieval}
\title{Internal for CombinatorialAggregate Coereba_MarkerExpressions}
\usage{
DataRetrieval(x, data, binary, Column)
}
\arguments{
\item{x}{The two fluorophores to quadrant by}

\item{data}{The data.frame of clusters vs individual specimens, with the ratio values}

\item{binary}{The data.frame of markers vs clusters, with 0 and 1 values}

\item{Column}{The label name for the respective quadrant}
}
\description{
Internal for CombinatorialAggregate Coereba_MarkerExpressions
}
