% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_MarkerExpressions.R
\name{.Internal_Aggregate}
\alias{.Internal_Aggregate}
\title{Internal for Marker Expressions}
\usage{
.Internal_Aggregate(
  x,
  panel,
  BinaryFile,
  OriginalData,
  Metadata,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  MyPanel = MyPanel,
  scalefactor,
  scalefactorlabel
)
}
\arguments{
\item{x}{Something}

\item{panel}{Something}

\item{BinaryFile}{Something}

\item{OriginalData}{Something}

\item{Metadata}{Something}

\item{myfactor}{Something}

\item{normality}{Something}

\item{shape_palette}{Something}

\item{fill_palette}{Something}

\item{MyPanel}{Something}

\item{scalefactor}{Something}

\item{scalefactorlabel}{Something}
}
\description{
Internal for Marker Expressions
}
\keyword{Internal}
