% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Heatmap.R
\name{.Internal_Heatmap}
\alias{.Internal_Heatmap}
\title{Internal for Utility_Heatmap}
\usage{
.Internal_Heatmap(binary, panel, ...)
}
\arguments{
\item{binary}{Something}

\item{panel}{Something}

\item{...}{Something}
}
\description{
Internal for Utility_Heatmap
}
\keyword{internal}
