% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GateCutoffs.R
\name{GateCutoffs}
\alias{GateCutoffs}
\title{Rough estimate of the location of the dichotomized gates.}
\usage{
GateCutoffs(x, columns, sample.name)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{columns}{Columns of interest, numeric.}

\item{sample.name}{Keyword for sample names (ex. "GUID")}
}
\description{
Rough estimate of the location of the dichotomized gates.
}
\examples{
NULL
}
