% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateGates.R
\name{InternalGateUpdate}
\alias{InternalGateUpdate}
\title{Internal for UpdateGates}
\usage{
InternalGateUpdate(x)
}
\arguments{
\item{x}{A path to a .csv file}
}
\description{
Internal for UpdateGates
}
\keyword{internal}
