% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModernGateCutoffs.R
\name{LocalMinima}
\alias{LocalMinima}
\title{Internal for Gate Cutoffs}
\usage{
LocalMinima(theX, theY, w, therepeats, alternatename, ...)
}
\arguments{
\item{theX}{Something}

\item{theY}{Something}

\item{w}{Something}

\item{therepeats}{Something}

\item{alternatename}{Something}

\item{...}{Something}
}
\description{
Internal for Gate Cutoffs
}
\keyword{internal}
