% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModernGateCutoffs.R
\name{ModernGateCutoffs}
\alias{ModernGateCutoffs}
\title{Improved Gate Estimates}
\usage{
ModernGateCutoffs(x, subset, sample.name, remove.strings)
}
\arguments{
\item{x}{A GatingSet object}

\item{subset}{Node level of interest}

\item{sample.name}{The keyword where names are stored}

\item{remove.strings}{Values to be removed from the name}
}
\value{
A data.frame with estimated gate cutoffs for every marker for every specimen
}
\description{
Improved Gate Estimates
}
\examples{
NULL

}
