% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiReach.R
\name{MultiReach}
\alias{MultiReach}
\title{Cluster your .fcs files by designated gates.}
\usage{
MultiReach(
  x,
  column,
  subsample = NULL,
  ratio = NULL,
  starter,
  sample.name,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  bins,
  cutoff = NULL,
  reference
)
}
\arguments{
\item{x}{A Gating Set Object (ex. gs or gs[[1]])}

\item{column}{Columns you want to cluster by, defined by numeric position}

\item{subsample}{Optional: Downsample to a certain number of cells}

\item{ratio}{Return values as a ratio of the total starting cells}

\item{starter}{Column to start split on (ex. "BV711")}

\item{sample.name}{The keyword your individual samples are identified by (ex. "GUID")}

\item{experiment}{A way to provide the experiment name directly (ex. "January2024Optimization")}

\item{experiment.name}{The keyword your experiment name is stored, when experiment = NULL. (ex. "$DATE")}

\item{condition}{A way to provide condition name directly (ex. "PMA")}

\item{condition.name}{The keyword your condition name is stored, when condition = NULL. (ex. "TUBENAME")}

\item{bins}{Number of bins the cells are plotted as in the ggplots (ex. 270)}

\item{cutoff}{The minimal number of cells in a cluster (ex. 3)}

\item{reference}{The imported .csv file containing the gating location for panels fluorophores.}
}
\description{
Cluster your .fcs files by designated gates.
}
\examples{
NULL
}
