% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_modified_utility.R
\name{my_modified_utility}
\alias{my_modified_utility}
\title{Downsample for cell population or marker of choice}
\usage{
my_modified_utility(x, sample, channel, cellNumber)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{sample}{A keyword denoting sample name (ex. "GUID")}

\item{channel}{A list of columns of interest to retrieve data from, numeric.}

\item{cellNumber}{Total number of cells to downsample to.}
}
\value{
Some additional value to edit
}
\description{
Downsample for cell population or marker of choice
}
\examples{
NULL
}
