% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wholesale.R
\name{my_wholesale}
\alias{my_wholesale}
\title{Downsamples a population of interest}
\usage{
my_wholesale(x, sample, cellNumber)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{sample}{Keyword denoting sample name (ex. "GUID")}

\item{cellNumber}{Desired number of cells to downsample to.}
}
\value{
Some additional value to edit
}
\description{
Downsamples a population of interest
}
\examples{
NULL
}
