% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSDs.R
\name{SampleSDs}
\alias{SampleSDs}
\title{Returns Standard Deviations for Normalized Detectors}
\usage{
SampleSDs(x, subsets, sample, experiment, condition)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{subsets}{A subset of interest, (ex. "lymph)}

\item{sample}{Keyword denoting the sample name (ex. "GUID")}

\item{experiment}{Keyword denoting the experiment name (ex. "GROUPNAME")}

\item{condition}{Keyword denoting the condition name (ex. "TUBENAME")}
}
\description{
Returns Standard Deviations for Normalized Detectors
}
\examples{
NULL
}
