% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyNbyN.R
\name{TidyNbyN}
\alias{TidyNbyN}
\title{An early version of Luciernaga's NbyN plot.}
\usage{
TidyNbyN(
  x,
  subsets,
  level,
  column,
  starter,
  sample.name,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  bins,
  clearance,
  outpath,
  sourcelocation,
  reference
)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{subsets}{A Gating Hierarchy node containing cells of interest.}

\item{level}{unclear}

\item{column}{A list of columns for exprs, superceeded}

\item{starter}{The y-column fluor to start with. (ex. "BV711-A")}

\item{sample.name}{Keyword that contains the sample names (ex. "GUID")}

\item{experiment}{Directly assign experiment (ex. "JAN2024")}

\item{experiment.name}{Keyword that contains the experiment name (ex. "GROUPNAME")}

\item{condition}{Directly assign condition (ex. "Control")}

\item{condition.name}{Keyword that contains the condition name (ex. "TUBENAME")}

\item{bins}{For plotting resolution, how many bins to divide the cells into}

\item{clearance}{Multiplied by quantile value to dictate the margin of the plots}

\item{outpath}{Location to send the .pdf}

\item{sourcelocation}{Location of the plotting file}

\item{reference}{A data.frame containing the location of the dividing lines per specimen.}
}
\description{
An early version of Luciernaga's NbyN plot.
}
\examples{
NULL
}
