% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyPlotThemPlots.R
\name{TidyPlotThemPlots}
\alias{TidyPlotThemPlots}
\title{Early version Luciernaga's single color plot functionality}
\usage{
TidyPlotThemPlots(
  x,
  subsets,
  column,
  sample.name,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  bins,
  clearance,
  outpath,
  sourcelocation
)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{subsets}{Your population of interest (ex. "root")}

\item{column}{A list of columns (superceeded)}

\item{sample.name}{Keyword for which samples name is stored (ex. "GUID")}

\item{experiment}{Directly assign experiment name (ex. "JAN2024")}

\item{experiment.name}{Keyword for which experiment name is stored (ex. "GROUPNAME")}

\item{condition}{Directly assign condition name (ex. "Control")}

\item{condition.name}{Keyword for which condition name is stored (ex. "TUBENAME")}

\item{bins}{Granularity of the plots by how many bins the cells are divided into}

\item{clearance}{A number by which the quantile value is multiplied to provide a margin}

\item{outpath}{Location to which to store the generated plots}

\item{sourcelocation}{Location to find the plot file.}
}
\description{
Early version Luciernaga's single color plot functionality
}
\examples{
NULL
}
