% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Aggregate.R
\name{Utility_Aggregate}
\alias{Utility_Aggregate}
\title{Returns aggregate plots comparing factor of interest for given markers}
\usage{
Utility_Aggregate(
  BinaryFile,
  OriginalData,
  markers,
  myfactor,
  starter,
  shape_palette,
  fill_palette,
  panel,
  scalefactor,
  scalefactorlabel,
  label,
  cex,
  size,
  ...
)
}
\arguments{
\item{BinaryFile}{The Heatmap file}

\item{OriginalData}{The Ready file}

\item{markers}{NULL or a list of Fluorophores desire to aggregate}

\item{myfactor}{The metadata factor in Original data you wish to compare}

\item{starter}{The starting name for the Coereba cuts, used to find the columns}

\item{shape_palette}{Desired shape for the points by factor}

\item{fill_palette}{Desired fill for the points by factor}

\item{panel}{A .csv or dataframe containing the Fluorophores and their Marker names}

\item{scalefactor}{Desired adjust to the axis}

\item{scalefactorlabel}{Desired name for the axis}

\item{label}{If a single plot, replaces the title with the following string}

\item{cex}{Desired spacing for the beeswarm}

\item{size}{Desired size for the points}

\item{...}{Additional arguments, feeds to Internal_Aggregate and Utility_Behemoth}
}
\value{
The generated ggplots.
}
\description{
Returns aggregate plots comparing factor of interest for given markers
}
\examples{
NULL
}
