% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_AssociationNotAutomaticallyCausation.R
\name{Utility_AssociationNotAutomaticallyCausation}
\alias{Utility_AssociationNotAutomaticallyCausation}
\title{Plot for Pearson/Spearman}
\usage{
Utility_AssociationNotAutomaticallyCausation(
  data,
  varX,
  varY,
  myfactor,
  mymethod
)
}
\arguments{
\item{data}{A data.frame object containing your data}

\item{varX}{The column of variables to plot on the X axis}

\item{varY}{The column of variables to plot on the y axis}

\item{myfactor}{The column containing factor you want to compare}

\item{mymethod}{Choice of "pearson" or "spearman"}
}
\description{
Plot for Pearson/Spearman
}
