% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Behemoth.R
\name{Utility_Behemoth}
\alias{Utility_Behemoth}
\title{Coereba's main function. Runs normality, runs stats, plots the results.}
\usage{
Utility_Behemoth(
  data,
  var,
  myfactor,
  normality = NULL,
  specifiedNormality = NULL,
  correction = "none",
  override = 0.05,
  shape_palette,
  fill_palette,
  cex,
  size,
  TheValue,
  ...
)
}
\arguments{
\item{data}{A data.frame object with metadata and data columns}

\item{var}{The column name for your variable of interest}

\item{myfactor}{The column name for your column containing your factor to group by}

\item{normality}{The Normality test to be applied, "dagostino" or "shapiro". Default NULL}

\item{specifiedNormality}{Default NULL leading to non-parametric, can switch by specifying
"parametric" or "nonparametric".}

\item{correction}{Multiple comparison correction argument, default is set at "none"}

\item{override}{Internal, default 0.05. Set to 0.99 to force pairwise comparison in anova/kw.}

\item{shape_palette}{Palette corresponding to factor levels, designating each's shape}

\item{fill_palette}{Palette corresponding to factor levels, designating each's fill}

\item{cex}{The width of the ggbeeswarm bin}

\item{size}{Size for the ggbeeswarm circles.}

\item{TheValue}{width of corral bin argument for beeswarm.}

\item{...}{Additional arguments}
}
\value{
Some additional value to edit
}
\description{
Coereba's main function. Runs normality, runs stats, plots the results.
}
\examples{
NULL

}
