% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Boxplot.R
\name{Utility_BoxPlot}
\alias{Utility_BoxPlot}
\title{Plot boxplot/beeswarm}
\usage{
Utility_BoxPlot(data, var, myfactor, shape_palette, fill_palette)
}
\arguments{
\item{data}{A data.frame of your data}

\item{var}{A column of data interest}

\item{myfactor}{A column containing the factor you want to compare between}

\item{shape_palette}{palette corresponding to the factor, specifying the shape}

\item{fill_palette}{palette corresponding to the factor, specifying the fill}
}
\description{
Plot boxplot/beeswarm
}
\examples{
NULL
}
