% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_GeomDensity.R
\name{Utility_GeomDensity}
\alias{Utility_GeomDensity}
\title{Returns distribution overlay of population of interest}
\usage{
Utility_GeomDensity(data, var, myfactor)
}
\arguments{
\item{data}{A data.frame object containing your data}

\item{var}{Column in the data.frame of specific data interested in}

\item{myfactor}{Column containing the factor you want to compare between.}
}
\description{
Returns distribution overlay of population of interest
}
\examples{
NULL
}
