% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Heatmap.R
\name{Utility_Heatmap}
\alias{Utility_Heatmap}
\title{Produces a Coereba Heatmap showing Coereba Clusters by Markers}
\usage{
Utility_Heatmap(binary, panel, export = FALSE, outpath = NULL, filename = NULL)
}
\arguments{
\item{binary}{A data.frame of markers by cluster}

\item{panel}{A csv or data.frame containing Fluorophore and Marker}

\item{export}{Whether to export as a .png, default is FALSE}

\item{outpath}{File.path to the desired storage location}

\item{filename}{File name for exported image}
}
\value{
A ggplot locally or to export location
}
\description{
Produces a Coereba Heatmap showing Coereba Clusters by Markers
}
\examples{

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")
panelData <- read.csv(panelPath, check.names=FALSE)
binaryPath <- file.path(File_Location, "HeatmapExample.csv")
binaryData <- read.csv(binaryPath, check.names=FALSE)

ThePlot <- Utility_Heatmap(binary=binaryData, panel=panelPath,
 export=FALSE, outpath=NULL, filename=NULL)

}
