% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Leviathan.R
\name{Utility_Leviathan}
\alias{Utility_Leviathan}
\title{Similar to Behemoth, drops the greatest value for all groups.}
\usage{
Utility_Leviathan(
  data1,
  var,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  switch,
  panel,
  scalefactor,
  scalefactorlabel
)
}
\arguments{
\item{data1}{A data.frame object}

\item{var}{A column of the data.frame containing variable of interest}

\item{myfactor}{A column of the data.frame containing the factor to compare}

\item{normality}{Choice of normality test, "dagostino" or "shapiro"}

\item{shape_palette}{Palette using names of the factor and corresponding
shape}

\item{fill_palette}{Palette using names of the factor and corresponding fill}

\item{switch}{Unclear}

\item{panel}{Unclear}

\item{scalefactor}{Multiply the ratio by this number for easier
interpretability}

\item{scalefactorlabel}{Relabel the axis with the cell type of interest.
(Ex. "Vd2 T cells")}
}
\description{
Similar to Behemoth, drops the greatest value for all groups.
}
\examples{
NULL
}
