% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_LongitudinalBehemoth.R
\name{Utility_LongitudinalBehemoth}
\alias{Utility_LongitudinalBehemoth}
\title{Scales up Utility_Behemoth to handle two factors on the x-axis, allowing for longitudinal plotting}
\usage{
Utility_LongitudinalBehemoth(
  data,
  Variable,
  GlobalFactor,
  LocalFactor,
  PositionDodgeWith = 0.8,
  DotSize = 3,
  DotSpacing = 2,
  shape_palette,
  fill_palette,
  savePlot = FALSE,
  outpath = NULL,
  filename = NULL,
  dpi = 600,
  width = 6,
  height = 3
)
}
\arguments{
\item{data}{The data.frame for plotting}

\item{Variable}{The column name of data you want visualized}

\item{GlobalFactor}{The column name that you want to use as the primary factor on the x-axis}

\item{LocalFactor}{The column name that you want to use as the secondary factor on the x-axis}

\item{DotSize}{Default 3, sets the beeswarm dot size}

\item{DotSpacing}{Default 2, dictates area the dots sprawl across}

\item{shape_palette}{A list dictating the shapes the LocalFactor values should correspond to}

\item{fill_palette}{A list dictating the fills the LocalFactor values should correspond to}

\item{savePlot}{Default FALSE, saves a .png file to the designated location}

\item{outpath}{Default NULL, provide a file.path to storage location for savePlot}

\item{filename}{Default NULL sets filename to LongitudinalBehemoth, alternatively provide
your own filename.}

\item{dpi}{Default 600, sets pixels for savePlot}

\item{width}{Default 6, sets width for savePlot}

\item{height}{Default 3, sets height for savePlot}

\item{PositionDodgeWidth}{Default 0.8, sets how far apart the beeswarm and boxplots are from
those of the other LocalFactor}
}
\description{
Scales up Utility_Behemoth to handle two factors on the x-axis, allowing for longitudinal plotting
}
