% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_MarkerExpressions.R
\name{Utility_MarkerExpressions}
\alias{Utility_MarkerExpressions}
\title{Generate overall marker expressions across Coereba clusters,}
\usage{
Utility_MarkerExpressions(
  BinaryFile,
  OriginalData,
  myfactor,
  starter,
  shape_palette,
  fill_palette,
  panel,
  scalefactor,
  scalefactorlabel,
  label,
  plot,
  savePlot,
  filename = NULL,
  cex,
  size,
  crossbar,
  XAxisLevels
)
}
\arguments{
\item{BinaryFile}{The Coereba file with 0 and 1 for Marker expression by cluster}

\item{OriginalData}{The data.frame with the individual specimen info}

\item{myfactor}{The factor you want the points to resemble}

\item{shape_palette}{Shape of your points by factor}

\item{fill_palette}{Fill of your points by factor}

\item{panel}{A .csv or data.frame of your Fluorophores and Markers}

\item{scalefactor}{Not implemented yet}

\item{scalefactorlabel}{Not implemented yet}

\item{label}{Plot Label}

\item{plot}{Whether to return a plot, FALSE returns data.frame}

\item{savePlot}{Whether to also save the plot}

\item{filename}{What file name to save it as (location not yet implemented)}

\item{cex}{How cramped to make the beeswarm columns.}

\item{size}{What size to make the beeswarm points}

\item{crossbar}{mean or median}

\item{XAxisLevels}{provide a list of final names to rearrange the x axis data.}
}
\value{
returns the data.frame or a ggplot plot.
}
\description{
Generate overall marker expressions across Coereba clusters,
}
\examples{
Not at this time
}
