% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Patchwork.R
\name{Utility_Patchwork}
\alias{Utility_Patchwork}
\title{Generate the plots into a single pdf}
\usage{
Utility_Patchwork(x, filename, outfolder, thecolumns, therows)
}
\arguments{
\item{x}{A list of plots}

\item{filename}{Name to save the .pdf as}

\item{outfolder}{Location to save the .pdf}

\item{thecolumns}{The number of columns per page}

\item{therows}{The number of rows per page}
}
\description{
Generate the plots into a single pdf
}
\examples{
NULL

}
