% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_QQplots.R
\name{Utility_QQplots}
\alias{Utility_QQplots}
\title{Generate QQ plots to assess normality}
\usage{
Utility_QQplots(data, var, myfactor)
}
\arguments{
\item{data}{A data.frame object}

\item{var}{The column containing your values}

\item{myfactor}{The column containing the factor designations that
you want to compare.}
}
\value{
Some additional value to edit
}
\description{
Generate QQ plots to assess normality
}
\examples{
NULL
}
