% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Querry.R
\name{Utility_Querry}
\alias{Utility_Querry}
\title{Detect changes a heatmap level below your target}
\usage{
Utility_Querry(
  BinaryFile,
  OriginalData,
  filename,
  outfolder,
  panel,
  starter,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  scalefactor,
  scalefactorlabel,
  ...
)
}
\arguments{
\item{BinaryFile}{Unclear}

\item{OriginalData}{Unclear}

\item{filename}{Name to give the results}

\item{outfolder}{Folder to which to output the results}

\item{panel}{Unclear}

\item{starter}{Unclear}

\item{myfactor}{Column containing factor designation that you want to
differentially compare}

\item{normality}{Normality test to be used, ex. "dagostino" or "sharpwilks"}

\item{shape_palette}{palette that list your factors of interest}

\item{fill_palette}{palette that list your factor of interest}

\item{scalefactor}{Number you are multiplying the ratio by to make more
interpretable (ex. 1000)}

\item{scalefactorlabel}{The cells of interest, ex. "Vd2 T cells"}
}
\value{
Some additional value to edit
}
\description{
Detect changes a heatmap level below your target
}
\examples{
NULL
}
