% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Reaggregate.R
\name{Utility_Reaggregate}
\alias{Utility_Reaggregate}
\title{Combine clusters and check}
\usage{
Utility_Reaggregate(
  BinaryFile,
  OriginalData,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  panel,
  scalefactor,
  scalefactorlabel,
  label
)
}
\arguments{
\item{BinaryFile}{NA}

\item{OriginalData}{NA}

\item{myfactor}{NA}

\item{normality}{NA}

\item{shape_palette}{NA}

\item{fill_palette}{NA}

\item{panel}{NA}

\item{scalefactor}{NA}

\item{scalefactorlabel}{NA}

\item{label}{NA}
}
\value{
NA
}
\description{
Combine clusters and check
}
\examples{
NA
}
