% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Rearrange.R
\name{Utility_Rearrange}
\alias{Utility_Rearrange}
\title{Rearrange FlowJo .csv files for nested questions}
\usage{
Utility_Rearrange(x, pattern, target)
}
\arguments{
\item{x}{the data.table}

\item{pattern}{recognizable string}

\item{target}{Level of nesting}
}
\value{
Some additional value to edit
}
\description{
Rearrange FlowJo .csv files for nested questions
}
\examples{
NULL
}
