% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Rooted.R
\name{Utility_Rooted}
\alias{Utility_Rooted}
\title{Take FlowJo .csv files and convert values to designated root.}
\usage{
Utility_Rooted(data, column)
}
\arguments{
\item{data}{A data.frame object}

\item{column}{A given column of the data.frame}
}
\value{
Some additional value to edit
}
\description{
Take FlowJo .csv files and convert values to designated root.
}
\examples{
NULL
}
