% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Stats.R
\name{Utility_Stats}
\alias{Utility_Stats}
\title{Normality Test followed by t-test/Anova}
\usage{
Utility_Stats(
  data,
  var,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  switch
)
}
\arguments{
\item{data}{A data.frame object containing the data values}

\item{var}{A data.frame column of interest}

\item{myfactor}{A data.frame column for which you want to differentionally compare if they are different}

\item{normality}{The Normality test to be applied, "dagostino" or "shapiro"}

\item{shape_palette}{provide the shape palette matches your provided factor
names}

\item{fill_palette}{provide the fill palette matches your provided factor
names}

\item{switch}{Unclear}
}
\description{
Normality Test followed by t-test/Anova
}
\examples{
NULL
}
