% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba-package.R
\docType{package}
\name{Coereba-package}
\alias{Coereba}
\alias{Coereba-package}
\title{Coereba: Dichotomized Boolean Clustering and Heatmaps for Spectral Flow Cytometry}
\description{
To learn more about how to use Coereba, please start by browsing the vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavidRach/Coereba}
  \item Report bugs at \url{https://github.com/DavidRach/Coereba/issues}
}

}
\author{
\strong{Maintainer}: David Rach \email{drach@som.umaryland.edu} (\href{https://orcid.org/0000-0003-2235-5585}{ORCID})

}
\keyword{internal}
