% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_Concatenate.R
\name{Coereba_Concatenate}
\alias{Coereba_Concatenate}
\title{Processes multiple Coereba flowframes, retrieves contents according to our Concatenate criteria, then cumulatively processes}
\usage{
Coereba_Concatenate(
  Set,
  metadata_cols,
  outpath = NULL,
  filename = "CombinedCoerebaFile"
)
}
\arguments{
\item{Set}{The flow/cyto set we are working with}

\item{metadata_cols}{The columns from pData to retain.}

\item{outpath}{Where to store the .fcs file}

\item{filename}{What to call the .fcs file}
}
\description{
Processes multiple Coereba flowframes, retrieves contents according to our Concatenate criteria, then cumulatively processes
}
\examples{
A <- 2+2
}
