% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_GateCutoffs.R
\name{Coereba_GateCutoffs}
\alias{Coereba_GateCutoffs}
\title{Function to guess the gate cutoff values}
\usage{
Coereba_GateCutoffs(
  gs,
  subset,
  sample.name,
  desiredCols = NULL,
  returnTemplate = FALSE,
  outpath = NULL,
  GatingTemplate = NULL,
  returnPlots = FALSE
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{subset}{Gate node of interest}

\item{sample.name}{The keyword where specimens name is stored}

\item{desiredCols}{Column names of fluorophores you want to gate}

\item{returnTemplate}{Default FALSE, when set to TRUE returns a .csv of Gating
Template that can be modified and provisioned by the GatingTemplate argument}

\item{outpath}{Default NULL, provides file path to desired folder to store returnTemplate}

\item{GatingTemplate}{A file.path to the GatingTemplate .csv you want to swap in}

\item{returnPlots}{Default FALSE, will instead return plots for troubleshooting}
}
\value{
A data.frame with estimated gate cutoffs for every marker
 for every specimen
}
\description{
Function to guess the gate cutoff values
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Coereba")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs", full.names = TRUE)
UnmixedFCSFiles <- FCS_Files[1]
UnmixedCytoSet <- load_cytoset_from_fcs(UnmixedFCSFiles,
 truncate_max_range = FALSE, transformation = FALSE)
UnmixedGatingSet <- GatingSet(UnmixedCytoSet)
Markers <- colnames(UnmixedCytoSet)
KeptMarkers <- Markers[-grep("Time|FS|SC|SS|Original|-W$|-H$|AF", Markers)]
biex_transform <- flowjo_biexp_trans(channelRange = 256, maxValue = 1000000,
 pos = 4.5, neg = 0, widthBasis = -1000)
TransformList <- transformerList(KeptMarkers, biex_transform)
flowWorkspace::transform(UnmixedGatingSet, TransformList)
UnmixedGates <- fread(file.path(path = File_Location,
 pattern = 'GatesUnmixed.csv'))
UnmixedGating <- gatingTemplate(UnmixedGates)
gt_gating(UnmixedGating, UnmixedGatingSet)

TheGateCutoffs <- Coereba_GateCutoffs(x=UnmixedGatingSet[1],
 subset="live", sample.name="GROUPNAME")

}
