% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_MarkerExpressions.R
\name{Coereba_MarkerExpressions}
\alias{Coereba_MarkerExpressions}
\title{Generate overall marker expressions across Coereba clusters,}
\usage{
Coereba_MarkerExpressions(
  data,
  binary,
  panel,
  starter,
  returnType = "All",
  CombinatorialArgs = NULL
)
}
\arguments{
\item{data}{The data.frame of clusters vs individual specimens, with the ratio values}

\item{binary}{The data.frame of markers vs clusters, with 0 and 1 values}

\item{panel}{A .csv or data.frame containing Fluorophore and Marker columns of your panel markers.}

\item{starter}{A string containing the starting characters for all Coereba cluster names}

\item{returnType}{Either "All" or "Combinatorial". Default is ALL.}

\item{CombinatorialArgs}{When returnType Combinatorial, the two fluorophores to create quadrants for.}
}
\value{
returns a data.frame
}
\description{
Generate overall marker expressions across Coereba clusters,
}
\examples{
NULL
}
