% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coereba_Processing.R
\name{Coereba_Processing}
\alias{Coereba_Processing}
\title{Takes Utility_Coereba output, plus additional metadata,and returns as a
 Bioconductor Summarized Experiment object}
\usage{
Coereba_Processing(
  x,
  metadata = NULL,
  metadata_columns = NULL,
  Identity = "specimen",
  metadataTemplate = FALSE,
  outpath = NULL
)
}
\arguments{
\item{x}{An object with Coereba gating and metadata. Acccepted arguments
include  a file.path to an .fcs, a flowFrame or cytoFrame, and a data.frame.}

\item{metadata}{A data.frame or file.path to a metadata.csv file}

\item{metadata_columns}{A vector of column names from the metadata want to incorporate.}

\item{Identity}{Column designating specimen in the metadata}

\item{metadataTemplate}{When metadata is NULL, returns a template data.frame containing specimen list.}

\item{outpath}{Default NULL, alternate a file.path to send metadataTemplate output to.}
}
\value{
A summarized experiment object
}
\description{
Takes Utility_Coereba output, plus additional metadata,and returns as a
 Bioconductor Summarized Experiment object
}
\examples{

library(Coereba)

}
