% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateGates.R
\name{UpdateGates}
\alias{UpdateGates}
\title{Update a gate cutoff csv with the CoerebaApp edits}
\usage{
UpdateGates(Clicks, Old, fileName, outpath)
}
\arguments{
\item{Clicks}{A path to a .csv file or a folder containing only the Click .csv files.}

\item{Old}{File path to the CSV containing the original Estimated Gate Cutoffs}

\item{fileName}{Desired name for the updated .csv file}

\item{outpath}{File path to save new .csv file at.}
}
\value{
An updated .csv file to the new location, and a data.frame
}
\description{
Update a gate cutoff csv with the CoerebaApp edits
}
\examples{
NULL
}
