% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Behemoth.R
\name{Utility_Behemoth}
\alias{Utility_Behemoth}
\title{Coereba's main function. Runs normality, runs stats, plots the results.}
\usage{
Utility_Behemoth(
  data,
  var,
  myfactor,
  normality = NULL,
  specified = NULL,
  correction = "none",
  Override = 0.05,
  shape_palette,
  fill_palette,
  cex,
  size,
  TheValue,
  ...
)
}
\arguments{
\item{data}{A data.frame of your data.}

\item{var}{A column containing your variables of interest.}

\item{myfactor}{A column containing the factor you want to compare}

\item{normality}{Choice of normality test, use NULL or "dagostino" or "shapiro"}

\item{specified}{In absence of normality test, dictates "parametric" or "nonparametric"}

\item{correction}{Choice of multiple choice correction, default is set at "none"}

\item{Override}{Not for general use, increase to 0.99 to force pairwise comparison past anova/kw.}

\item{shape_palette}{Palette corresponding to factor levels, designating each's shape}

\item{fill_palette}{Palette corresponding to factor levels, designating each's fill}

\item{cex}{The width of the ggbeeswarm bin}

\item{size}{Size for the ggbeeswarm circles.}

\item{TheValue}{width of corral bin argument for beeswarm.}

\item{...}{Additional arguments}
}
\value{
Some additional value to edit
}
\description{
Coereba's main function. Runs normality, runs stats, plots the results.
}
\examples{
NULL

}
