% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Behemoth.R
\name{Utility_Behemoth}
\alias{Utility_Behemoth}
\title{Coereba's main function. Runs normality, runs stats, plots the results.}
\usage{
Utility_Behemoth(
  data,
  var,
  myfactor,
  normality,
  shape_palette,
  fill_palette,
  switch
)
}
\arguments{
\item{data}{A data.frame of your data.}

\item{var}{A column containing your variables of interest.}

\item{myfactor}{A column containing the factor you want to compare}

\item{normality}{Choice of normality test, "dagostino" or "shapiro"}

\item{shape_palette}{Palette corresponding to factor levels, designating each's shape}

\item{fill_palette}{Palette corresponding to factor levels, designating each's fill}

\item{switch}{Unclear}
}
\value{
Some additional value to edit
}
\description{
Coereba's main function. Runs normality, runs stats, plots the results.
}
\examples{
NULL

}
