% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Behemoth.R
\name{Utility_Behemoth}
\alias{Utility_Behemoth}
\title{From a data.frame, runs stats on column of interest, and returns the ggplot for it.}
\usage{
Utility_Behemoth(
  data,
  var,
  myfactor,
  normality = NULL,
  specifiedNormality = NULL,
  correction = "none",
  override = 0.05,
  shape_palette,
  fill_palette,
  cex = 2,
  size = 3,
  corral.width = 1,
  XAxisLevels = NULL,
  statLines = TRUE,
  statsHeight = NULL
)
}
\arguments{
\item{data}{A data.frame object with metadata and data columns}

\item{var}{The column name for your variable of interest}

\item{myfactor}{The column name for your column containing your factor to group by}

\item{normality}{The Normality test to be applied, "dagostino" or "shapiro". Default NULL}

\item{specifiedNormality}{Default NULL leading to non-parametric, can switch by specifying
"parametric" or "nonparametric".}

\item{correction}{Multiple comparison correction argument, default is set at "none"}

\item{override}{Internal, default 0.05. Set to 0.99 to force pairwise comparison in anova/kw.}

\item{shape_palette}{Palette corresponding to factor levels, designating each's shape}

\item{fill_palette}{Palette corresponding to factor levels, designating each's fill}

\item{cex}{The width of the ggbeeswarm bin}

\item{size}{Size for the ggbeeswarm circles.}

\item{corral.width}{width of corral bin argument for beeswarm.}

\item{XAxisLevels}{Provide list marker names correct order for x-axis reordering, default NULL}

\item{statLines}{Default is TRUE, otherwise skips plotting pvalue and brackets}

\item{statsHeight}{Default is NULL, when provided, sets stat line y-axis height.}
}
\value{
A ggplot2 object for the corresponding data of interest.
}
\description{
From a data.frame, runs stats on column of interest, and returns the ggplot for it.
}
\examples{

shape_ptype <- c("HU" = 22, "HEU-lo" = 21, "HEU-hi" = 21)
fill_ptype <- c("HU" = "white", "HEU-lo" = "darkgray", "HEU-hi" = "black")

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")
binaryPath <- file.path(File_Location, "HeatmapExample.csv")
dataPath <- file.path(File_Location, "ReadyFileExample.csv")
panelData <- read.csv(panelPath, check.names=FALSE)
binaryData <- read.csv(binaryPath, check.names=FALSE)
dataData <- read.csv(dataPath, check.names=FALSE)

All <- Coereba_MarkerExpressions(data=dataData, binary=binaryData,
 panel=panelData, starter="SparkBlue550")

Plot <- Utility_Behemoth(data=All, var="CD62L", myfactor="ptype",
 normality="dagostino", correction="none", shape_palette=shape_ptype,
 fill_palette=fill_ptype, XAxisLevels = c("HU", "HEU-lo", "HEU-hi"))

}
