% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Coereba.R
\name{Utility_Coereba}
\alias{Utility_Coereba}
\title{Runs Coereba for Dichotomized Gating Annotation}
\usage{
Utility_Coereba(
  x,
  subsets,
  sample.name,
  subsample = NULL,
  columns = NULL,
  notcolumns = NULL,
  reference,
  starter,
  inverse.transform = TRUE
)
}
\arguments{
\item{x}{A Gating Set object}

\item{subsets}{The desired GatingHierarchy subset}

\item{sample.name}{Keyword for sample name}

\item{subsample}{An optional downsample, doesn't work with inverse.transform=TRUE}

\item{columns}{An optional way to select columns to keep.}

\item{notcolumns}{An optional way to remove select columns.}

\item{reference}{The external data.frame or path to the .csv with the
specified gate split points by specimen and marker.}

\item{starter}{The column name to start the splits with}

\item{inverse.transform}{Whether to reverse the data transform after Coereba cluster
is calculated, Default is set to TRUE to allow for .fcs export.}
}
\value{
A data.frame
}
\description{
Runs Coereba for Dichotomized Gating Annotation
}
\examples{
NULL
}
