% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Coereba.R
\name{Utility_Coereba}
\alias{Utility_Coereba}
\title{Run Coereba and return as a data.frame}
\usage{
Utility_Coereba(
  x,
  subsets,
  sample.name,
  subsample = NULL,
  columns,
  notcolumns,
  reference,
  starter,
  ReturnType = "data"
)
}
\arguments{
\item{x}{A Gating Set object}

\item{subsets}{The desired GatingHierarchy subset}

\item{sample.name}{Keyword for sample name}

\item{subsample}{If desired, downsample to a certain number}

\item{columns}{Columns to Keep.}

\item{notcolumns}{Columns to Remove}

\item{reference}{The external data.frame with the specified gate cutoffs}

\item{starter}{The column name to start the splits with}

\item{ReturnType}{Whether to return transformed (default) or inverse.transformed data ("inversed")}
}
\value{
A data.frame
}
\description{
Run Coereba and return as a data.frame
}
\examples{
Not at this time
}
