% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Heatmap.R
\name{Utility_Heatmap}
\alias{Utility_Heatmap}
\title{Returns a Coereba Heatmap}
\usage{
Utility_Heatmap(cells, filename, return)
}
\arguments{
\item{cells}{Unclear, Coereba output including Cluster column}

\item{filename}{Name to save .png file as}
}
\description{
Returns a Coereba Heatmap
}
\examples{
NULL
}
