% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Heatmap.R
\name{Utility_Heatmap}
\alias{Utility_Heatmap}
\title{Returns a Coereba Heatmap}
\usage{
Utility_Heatmap(cells, filename, return, panel)
}
\arguments{
\item{cells}{Unclear, Coereba output including Cluster column}

\item{filename}{Name to save .png file as}

\item{return}{Whether to save the file as a png, TRUE/FALSE}

\item{panel}{A csv file containing a panel}
}
\value{
Some additional value to edit
}
\description{
Returns a Coereba Heatmap
}
\examples{
NULL
}
