% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Reaggregate.R
\name{Utility_Reaggregate}
\alias{Utility_Reaggregate}
\title{Combine clusters and check}
\usage{
Utility_Reaggregate(
  BinaryFile,
  OriginalData,
  myfactor,
  starter,
  normality,
  shape_palette,
  fill_palette,
  panel,
  scalefactor,
  scalefactorlabel,
  label,
  ...
)
}
\arguments{
\item{BinaryFile}{The Heatmap file}

\item{OriginalData}{The Ready file}

\item{myfactor}{The metadata factor in Original data you wish to compare}

\item{starter}{First parameter Coereba split by, used to identify columns that aren't metadata}

\item{normality}{Choice of normality test, "dagostino" or "shapiro"}

\item{shape_palette}{Desired shape for the points by factor}

\item{fill_palette}{Desired fill for the points by factor}

\item{panel}{A .csv or dataframe containing the Fluorophores and their Marker names}

\item{scalefactor}{Desired adjust to the axis}

\item{scalefactorlabel}{Desired name for the axis}

\item{label}{Replaces the title with the following string}
}
\value{
NA
}
\description{
Combine clusters and check
}
\examples{
NA
}
