% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Stats.R
\name{Utility_Stats}
\alias{Utility_Stats}
\title{Normality Test followed by t-test/Anova}
\usage{
Utility_Stats(
  data,
  var,
  myfactor,
  normality = NULL,
  specifiedNormality = NULL,
  correction = "none",
  override = 0.05,
  returnType = "stats"
)
}
\arguments{
\item{data}{A data.frame object with metadata and data columns}

\item{var}{The column name for your variable of interest}

\item{myfactor}{The column name for your column containing your factor to group by}

\item{normality}{The Normality test to be applied, "dagostino" or "shapiro". Default NULL}

\item{specifiedNormality}{Default NULL leading to non-parametric, can switch by specifying
"parametric" or "nonparametric".}

\item{correction}{Multiple comparison correction argument, default is set at "none"}

\item{override}{Internal, default 0.05. Set to 0.99 to force pairwise comparison in anova/kw.}

\item{returnType}{Internal, default is "stats". "behemoth" for internal usage.}
}
\value{
A data.frame of test results
}
\description{
Normality Test followed by t-test/Anova
}
\examples{

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")
binaryPath <- file.path(File_Location, "HeatmapExample.csv")
dataPath <- file.path(File_Location, "ReadyFileExample.csv")
panelData <- read.csv(panelPath, check.names=FALSE)
binaryData <- read.csv(binaryPath, check.names=FALSE)
dataData <- read.csv(dataPath, check.names=FALSE)

All <- Coereba_MarkerExpressions(data=dataData, binary=binaryData,
 panel=panelData, starter="SparkBlue550")

TheStats <- Utility_Stats(data=All, var="CD62L",
 myfactor="ptype", normality="dagostino")

}
